defer the 'press enter to exit' prompt
On Windows, the user is prompted to press the enter key to exit the program. This happens before z64rom has finished, so if the user terminates the program any other way (e.g. by clicking the 'X' on the terminal window), problems can arise on subsequent execution because some changes were not saved. This change defers the exit prompt until it's safe to display it. Only display the exit prompt if the program has no further operations pending.