patch and unpatch instant scene load hack based on previous run
There was an issue with subsequent runs following the use of the --instant feature. Running with --instant would always trigger a rebuild for certain compilation units and patch the game's boot process so it loads into a level instantly, but rerunning the program afterwards without --instant would not undo the patch. As of this commit, the program now keeps track of whether or not --instant was provided on the previous run, and if it was, what its parameters were. If the current run differs from the previous run, a rebuild is triggered so the game will either be patched again with the updated settings, or unpatched accordingly.