ensure actor id 0 is never dumped nor injected
Fixes a corner case that allowed actor 0 to be dumped, causing a crash during game boot following subsequent builds. Also includes a safety check that throws a helpful error message while building if user tries to overwrite actor 0. Actor 0 is reserved for the player actor and should not be used. Thank you for the bug report and for testing @Zeldaboy14