Fixs for upped object count
Disabled the Fault_AddClient (can't seem to get the Actor_FaultPrint to hook). Moved Actor_SetObjectDependency to lib_user, as lib_code crashs if you replace the code there. It was also deleted from lib_code too. Player_DrawHookshotReticle is also updated to include the headers for link, specifically for the Hookshot's Reticle. And Scene_CommandSpecialFiles has been updated to include the proper h files for sNaviMsgFiles. Thanks @Nokaubure for helping with this last-minute issues this morning!