delete bad macro override
The substituted macro was identical to the original in z64hdr, except it was not gluing the suffix _BYTES onto typenames, This was causing problems rendering the recompiled title screen. This should be fairly backwards-compatible: any code relying on the override will cause compiler errors to be thrown so users know to update their code.