feature: build-time scene embeds & scene-func compiler
If a scene function is included in the same folder as a zscene, it will be compiled into a relocatable overlay that gets embedded into the scene while building your rom. Your original scene will not be modified. This will be a handy way to run scene-specific code such as animated textures. Additionally, I can imagine build-time embeds being useful for a minimap system in the future.